Refactor: move get_minio_admin to utils
